# openapi.yml
...
openapi: 3.1.0
info:
  title: Postmill API
  version: 3.4.0
  description: |
    Comprehensive OpenAPI description of the **Postmill** backend (NestJS) as it exists in
    `postmill-ai/postmill-app`.

    Postmill schedules social-media and chat posts to 28+ channels, with a calendar, analytics, team
    management, a media library, and — in this fork — a pluggable multi-provider **AI layer**,
    **Analytics v2** persisted dashboards, **social comments** sync/moderation, admin-configurable
    **channel OAuth credentials**, and an OAuth2 authorization-code provider.

    ## Surfaces & authentication

    There are three distinct authentication surfaces:

    1. **Dashboard API** (the bulk of routes). Requires a Postmill user session JWT supplied either as
       the `auth` **cookie** or the `auth` **request header**. The active organization is selected
       with the `showorg` cookie/header. Many routes are further gated by RBAC policies
       (`@CheckPolicies`) or by `isSuperAdmin`.
    2. **Public integration API** (`/public/v1/*`). Requires an organization **API key** (or a
       `pos_`-prefixed OAuth access token) in the `Authorization` header. Intended for
       n8n/Zapier/CLI/3rd-party automation.
    3. **Unauthenticated routes** — registration/login, the OAuth2 authorization/token endpoints,
       Stripe & enterprise webhooks, the social-connect OAuth callback, health/monitor, and public
...

...
    > global route prefix — controller paths are mounted at the root.

...

...
servers:
...
security:
  - cookieAuth: []
  - authHeader: []
paths:
...
  /media/designs/render:
    post:
      tags: [Media]
      summary: Server-side render a design document to PNG or PDF
      operationId: renderDesign
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [doc], properties: { doc: { type: object }, pageIndex: { type: integer }, format: { type: string, enum: [png, pdf] }, pixelRatio: { type: number }, transparent: { type: boolean } } } } }
      responses: { '200': { description: Rendered image/pdf, content: { image/png: { schema: { type: string, format: binary } }, application/pdf: { schema: { type: string, format: binary } } } } }
  /media/designs/bulk-generate:
    post:
      tags: [Media]
      summary: Data-driven bulk design generation ({{variables}} per row)
      operationId: bulkGenerateDesigns
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [doc, rows], properties: { doc: { type: object }, rows: { type: array, items: { type: object, additionalProperties: { type: string } } }, format: { type: string } } } } }
      responses: { '200': { description: Base64 image data URLs, content: { application/json: { schema: { type: object, properties: { images: { type: array, items: { type: string } }, truncated: { type: boolean }, totalRows: { type: integer } } } } } } }
...
  /settings/team:
    get:
      tags: [Settings]
      summary: List team members
      operationId: getTeam
      responses: { '200': { description: Team, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } } }
    post:
      tags: [Settings]
      summary: Invite a team member
      operationId: inviteTeamMember
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AddTeamMemberDto' } } }
      responses: { '200': { description: Invited } }
  /settings/team/{id}:
    delete:
      tags: [Settings]
      summary: Remove a team member
      operationId: deleteTeamMember
      parameters: [{ $ref: '#/components/parameters/PathId' }]
      responses: { '200': { description: Removed } }
  /settings/shortlink:
    get:
      tags: [Settings]
      summary: Get shortlink preference
      operationId: getShortlinkPreference
      responses: { '200': { description: Preference } }
    post:
      tags: [Settings]
      summary: Update shortlink preference (admin)
      operationId: updateShortlinkPreference
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ShortlinkPreferenceDto' } } }
      responses: { '200': { description: Updated } }

  # ===================================================================== BILLING
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Session JWT in the `auth` cookie (dashboard API).
    authHeader:
      type: apiKey
      in: header
      name: auth
      description: Session JWT in the `auth` header (alternative to the cookie).
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Organization **API key** or a `pos_`-prefixed OAuth access token, sent **raw** (no `Bearer`
        prefix) in the `Authorization` header. Used by `/public/v1/*`.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    AddTeamMemberDto:
      type: object
      required: [role, sendEmail]
      properties:
        email: { type: string, format: email, description: Required when sendEmail is true }
        role: { type: string, enum: [USER, ADMIN] }
        sendEmail: { type: boolean }
    ShortlinkPreferenceDto:
      type: object
      required: [shortlink]
      properties:
        shortlink: { type: string, description: ShortLinkPreference enum value }
...